Add files via upload
Hi Tetiana,

I wanted to update you regarding the conceptual model. I wasn’t able to simplify it further due to the technical limitations of the tool I was using. However, I fully understand that the conceptual model should focus only on the list of entities and the relationships between them.

If needed, I am reaady to remake or simplify it.